增加 商品分类管理-修改分类;删除分类 Add     删除分类相关(有子分类不允许删除) Modi    修改分类时[类型模板ID]的展示问题;0

+++ pinyougou-sellergoods-service/src/main/java/com/pinyougou/sellergoods/service/impl/ItemCatServiceImpl.java
@@ -9,6 +9,7 @@ import com.pinyougou.pojo.TbItemCatExample;
 import com.pinyougou.pojo.TbItemCatExample.Criteria;
 import com.pinyougou.sellergoods.service.ItemCatService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 import priv.zhong.bean.PageResult;
 
 import java.util.List;
@@ -19,6 +20,7 @@ import java.util.List;
  * @author Administrator
  */
 @Service
+@Transactional
 public class ItemCatServiceImpl implements ItemCatService {
 
     @Autowired
@@ -76,6 +78,13 @@ public class ItemCatServiceImpl implements ItemCatService {
     @Override
     public void delete(Long[] ids) {
         for (Long id : ids) {
+            TbItemCatExample example=new TbItemCatExample();
+            Criteria criteria = example.createCriteria();
+            criteria.andParentIdEqualTo(id);
+            List<TbItemCat> tbItemCats = itemCatMapper.selectByExample(example);
+            if(tbItemCats!=null &&tbItemCats.size()>0){
+                throw new RuntimeException("该分类有子分类,不能删");
+            }
             itemCatMapper.deleteByPrimaryKey(id);
         }
     }
